+++ src/backend/main/py.main/caliopen_main/message/objects/message.py
@@ -24,7 +24,6 @@ from ..parameters.draft import Draft
 from ..core import RawMessage
 from .attachment import MessageAttachment
 from .external_references import ExternalReferences
-from caliopen_main.user.objects.identities import Identity
 from .participant import Participant
 from schematics.types import UUIDType
 from caliopen_main.message.parameters.participant import \
@@ -72,7 +71,6 @@ class Message(ObjectIndexable):
         'date_sort': datetime.datetime,
         'discussion_id': UUID,
         'external_references': ExternalReferences,
-        'identities': [Identity],
         'importance_level': types.IntType,
         'is_answered': types.BooleanType,
         'is_draft': types.BooleanType,
@@ -88,6 +86,7 @@ class Message(ObjectIndexable):
         'tags': [types.StringType],
         'type': types.StringType,
         'user_id': UUID,
+        'user_identities': [UUID],
     }
 
     _json_model = ParamMessage
@@ -125,11 +124,11 @@ class Message(ObjectIndexable):
         if not strict_patch:
             allowed_properties = [
                 "body",
-                "identities",
                 "message_id",
                 "parent_id",
                 "participants",
                 "subject",
+                "user_identities",
             ]
             for key, value in params.items():
                 if key not in allowed_properties:
@@ -184,7 +183,7 @@ class Message(ObjectIndexable):
             allowed_properties = [
                 "body",
                 "current_state",
-                "identities",
+                "user_identities",
                 "message_id",
                 "parent_id",
                 "participants",
@@ -239,8 +238,8 @@ class Message(ObjectIndexable):
             discussion_id = find_or_create_discussion(user, params)
             self.discussion_id = discussion_id
 
-        if "identities" not in params and self.identities:
-            draft_param.identities = self_dict["identities"]
+        if "user_identities" not in params and self.user_identities:
+            draft_param.user_identities = self_dict["user_identities"]
 
         try:
             draft_param.validate_consistency(str(self.user_id), False)